feat: Implement database client management and API endpoints for MongoDB, Redis, and Qdrant
